requirements.txt
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0
xformers>=0.0.23
torch>=2.1.0
torchvision>=0.16.0

# Evaluation
openai-clip>=1.0.1
scikit-image>=0.21.0

# Data & Visualization
Pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0

# Bonus
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Utilities
huggingface_hub>=0.20.0
tqdm>=4.65.0
